on a roll, safety save

+++ src/main/java/com/codeup/adlister/dao/MySQLUsersDao.java
@@ -22,7 +22,7 @@ public class MySQLUsersDao implements Users {
     }
 
     @Override
-    public Long insert(User user) {
+    public int insert(User user) {
         String query = "INSERT INTO users(name, username, email, password) VALUES (?, ?, ?, ?)";
         try {
             PreparedStatement stmt = connection.prepareStatement(query, Statement.RETURN_GENERATED_KEYS);
@@ -33,7 +33,7 @@ public class MySQLUsersDao implements Users {
             stmt.executeUpdate();
             ResultSet rs = stmt.getGeneratedKeys();
             rs.next();
-            return rs.getLong(1);
+            return rs.getInt(1);
         } catch (SQLException e) {
             throw new RuntimeException("Error creating new user", e);
         }
@@ -68,7 +68,7 @@ public class MySQLUsersDao implements Users {
             return null;
         }
         return new User(
-            rs.getLong("id"),
+            rs.getInt("id"),
             rs.getString("name"),
             rs.getString("username"),
             rs.getString("email"),
