+++ src/main/java/com/codeup/adlister/models/Breed.java
@@ -1,19 +1,19 @@
 package com.codeup.adlister.models;
 
 public class Breed {
-    private long id;
+    private int id;
     private String name;
 
-    public Breed(long id, String name){
+    public Breed(int id, String name){
         this.id = id;
         this.name = name;
     }
 
-    public long getId() {
+    public int getId() {
         return id;
     }
 
-    public void setId(long id) {
+    public void setId(int id) {
         this.id = id;
     }
