+++ src/main/java/com/codeup/adlister/models/User.java
@@ -1,7 +1,7 @@
 package com.codeup.adlister.models;
 
 public class User {
-    private long id;
+    private int id;
     private String name;
     private String username;
     private String email;
@@ -16,7 +16,7 @@ public class User {
         this.password = password;
     }
 
-    public User(long id, String name, String username, String email, String password) {
+    public User(int id, String name, String username, String email, String password) {
         this.id = id;
         this.name = name;
         this.username = username;
@@ -24,11 +24,11 @@ public class User {
         this.password = password;
     }
 
-    public long getId() {
+    public int getId() {
         return id;
     }
 
-    public void setId(long id) {
+    public void setId(int id) {
         this.id = id;
     }
